+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 ---
+###
+---
+### 21 DEC 2024
+- Created `itemsRouter.js` module.
+- Consolidated add/edit/delete routers and controllers with their respective parent path. For example, moved `getAddCategory` from the `addController.js` module to the `categoriesController.js` module.
+- Deleted the following modules: `addController`, `editController`, `deleteController`, `addRouter`, `editRouter`, and `deleteRouter`.
+- Deleted `success` page and created a `success` partial.
+- All `POST` requests now redirect to specific pages on successful validations. For example, `postEditCategory` will  redirect to `/categories`, and `postAddItem` will redirect to `/item/add`.
+---
 ### 20 DEC 2024
 - Inserted logical OR (`||`) operator in `pool.query(text: string, values?: any[])` for quantity and price values during the `insertItem` query. If the quantity and price values are empty strings, then their values will be set to `null`.
-- Successfully submitting a new category will call a redirect to `/add/category/?success=true`.
+- Currently, submitting a new category will call a redirect to `/add/category/?success=true`.
 ---
 ### 19 DEC 2024
 - Items can be edited when the form is successfully submitted.